Implement cancel + playback complete flushing
The old code's cancel method did not comply with the datasheet for the
vs1053 when completeing or canceling an audio stream.

The vs1053 requires flushing the input buffer with a series of fill
bytes. This code implements sections 9.5.1 an 9.5.2 of the vs1053
datasheet which describe how to correctly complete a playback stream
and cancel playback.